Misc
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@49488 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/exe.h
@@ -193,7 +193,7 @@ public:
 	dsc desc;
 	USHORT length;
 	bool intl;
-	unsigned impure;
+	ULONG impure;
 	Firebird::HalfStaticArray<sort_key_def, 2> keyItems;
 };
